refactor: make enactTogglePasswordView be module
