+++ public/index.html
@@ -89,7 +89,6 @@
   <script src="js/utils/get-username.js"></script>
   <script src="js/services/is-logged-in.js"></script>
   <script type="module" src="js/app.js"></script>
-  <script src="js/utils/toggle-password-view.js"></script>
   <script src="js/utils/toggle-sign-mode.js"></script>
   <script src="js/utils/make-query-strings.js"></script>
   <script src="js/services/sign-up.js"></script>
